# openapi/instrument-service.yaml
...
openapi: 3.1.0
info:
  title: ZeroSum instrument-service API
  version: "1.0.0"
  description: |
    Payment attempts and the providers behind them.

    **Attempt history is the audit trail.** A status is never overwritten in place: every change appends a row to the
    attempt's transition history with the cause that produced it, so "why is this attempt in `NEEDS_REVIEW`" is
    answered by reading rows rather than by correlating logs. The read operation returns that history in sequence
    order.

    **What is not here.** The instrument token is never returned: it is the one field on an attempt that identifies a
    payment instrument, and an endpoint that returned it would put it in every client log (D00-8).
servers:
  - url: http://localhost:8083
    description: Local Compose topology
security:
  - bearerAuth: []
paths:
...
  /v1/payment-attempts:
    get:
      operationId: listRecentPaymentAttempts
      summary: The newest attempts first
      description: |
        Requires the **reader** role. The same attempt document as the single read, without its history, newest
        first. `kind` narrows the list to charges, refunds or payouts. A `limit` outside 1..500 or an unknown `kind`
        is refused with `invalid_query` rather than clamped or ignored.
      tags: [payment-attempts]
      parameters:
        - name: kind
          in: query
          required: false
          schema: { type: string, enum: [CHARGE, REFUND, PAYOUT] }
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Up to `limit` attempts, newest first.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Attempt' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/Unavailable' }

  # decision: S09 — the operator dashboard (docs/results/s09/dashboard.md)
  /v1/payment-attempts/summary:
    get:
      operationId: getPaymentAttemptsSummary
      summary: How many attempts are in each status
      description: |
        Requires the **reader** role. Every attempt counted by kind, provider, status and currency, with the money
        they carry, from one grouped statement.
      tags: [payment-attempts]
      responses:
        '200':
          description: The totals.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AttemptsSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /v1/payment-attempts/{attempt_id}:
    get:
      operationId: getPaymentAttempt
      summary: One attempt with its transition history
      description: |
        Requires the **reader** role, which writer and admin tokens also satisfy.

        Returns the attempt as it stands plus every status it has been through, in `seq` order. History is
        append-only and numbered per attempt, so the sequence is the order the statuses actually happened in.
      tags: [payment-attempts]
      parameters:
        - $ref: '#/components/parameters/AttemptId'
      responses:
        '200':
          description: The attempt and its history.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AttemptDetail' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /v1/payment-attempts/{attempt_id}/cancel:
    post:
      operationId: cancelPaymentAttempt
      summary: Cancel an attempt that has not been submitted
      description: |
        Requires the **admin** role: a cancellation changes money state, so it is not a reader operation.

        Cancellation is allowed **only from `CREATED`** (master §5.11). Anything else is a `not_cancellable` conflict
        whose detail names the status the attempt is actually in — the caller's next step differs, since an attempt
        already `SUBMITTING` may still decline on its own, while one already `SUCCEEDED` has moved money and needs a
        refund rather than a cancellation.

        A cancel racing a submit is settled by the optimistic version guard rather than by the status check above:
        exactly one of them wins, and the loser gets the same conflict with the current status.
      tags: [payment-attempts]
      parameters:
        - $ref: '#/components/parameters/AttemptId'
      responses:
        '200':
          description: Cancelled. The body is the attempt, now `CANCELLED`, with the cancellation in its history.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AttemptDetail' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
  /v1/payout-runs:
    # decision: S09 — the operator dashboard (docs/results/s09/dashboard.md)
    get:
      operationId: listPayoutRuns
      summary: The newest payout runs first
      description: |
        Requires the **reader** role: listing runs moves no money. Refused runs are included, because a refusal is
        a recorded run and the evidence of why nobody was paid. Per-driver results are folded into `outcomes` and
        `paid_minor` rather than returned whole.
      tags: [payouts]
      parameters:
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Up to `limit` runs, newest first.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PayoutRunListing' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/Unavailable' }
    post:
      operationId: createPayoutRun
      summary: Pay every eligible driver what the ledger says they are owed
      description: |
        Requires the **writer** role, and an `Idempotency-Key`: this is the operation that sends money out, and a
        retried POST without a key would pay every driver twice. The stored run **is** the idempotency record
        (D03-3) — the same key with the same body replays it, the same key with a different body is
        `idempotency_key_reused`.

        **One run, one currency.** A driver owed money in two currencies is paid by two runs, which is what makes
        the amount unambiguous and matches the one-in-flight-payout rule below being per (driver, currency).

        **Order of checks** (D05-7): payouts kill switch → pipeline freshness → per-driver eligibility. A driver is
        eligible when they have a registered instrument at a provider that can disburse, their `payable` balance on
        its normal side is at least the configured minimum, and they have no payout already in flight. The **amount
        is that balance, read from the ledger at that moment** (ADR-0004), and the `as_of_seq` it was read at is
        recorded against the driver's result for audit.

        **At most one in-flight payout per (driver, currency)** — M10(a) — is enforced by the `one_inflight_payout`
        partial unique index, not by a check in this service. Two runs racing for one driver therefore both proceed;
        one inserts and the other is told by the database that it lost, and reports that driver as
        `SKIPPED_IN_FLIGHT`. A losing run is a normal `201`, never an error.

        **Refusals are `409` and are recorded.** `ledger_stale` (M10(c)) means the pipeline staleness — order-service
        outbox age **plus** ledger unapplied age **plus** this service's outbox age — exceeded the 5 s threshold, *or*
        that one of those stages could not be measured at all. Those are one code on purpose: a stage we could not
        read is not a stage we may assume is healthy, and the alternative is paying against balances nobody checked.
        `payouts_disabled` means the kill switch is off. **A refused run creates no payout attempts**, and because
        the refusal is stored under the key, retrying with the *same* key replays the refusal — retry with a fresh
        key once the pipeline has caught up.

        **Accepted is not settled.** A created attempt is submitted to the bank after this response is prepared and
        reaches `PENDING`, emitting `PAYOUT_ACCEPTED`. Settlement and any return arrive later by webhook.
      tags: [payouts]
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, maxLength: 255 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PayoutRunRequest' }
      responses:
        '201':
          description: The run completed. Every driver it considered is in `results` with the outcome and why.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PayoutRun' }
        '200':
          # Quoted: an unquoted `key: value` inside a plain scalar reads as a nested mapping and fails to parse.
          description: 'A replay of a run that already exists, carrying the `Idempotent-Replayed` header.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PayoutRun' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: >
            Refused, and no payout attempt was created: `ledger_stale` when the pipeline is behind the 5 s threshold
            or could not be measured, `payouts_disabled` when the kill switch is off. The refusal is recorded as a
            run, whose id is in the detail.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422':
          description: The key was reused with a different request, or the currency is outside the allow-list.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Static per-role bearer token (master §5.11), verified by the shared `libs/auth` module. Reading an attempt
        needs the reader role; cancelling one needs admin.
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      description: Page size. Absent means 50; anything but an integer in 1..500 is refused with `invalid_query`.
      schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
    AttemptId:
      name: attempt_id
      in: path
      required: true
      description: The attempt's UUID, which is also the client reference sent to the provider.
      schema: { $ref: '#/components/schemas/Uuid' }
  responses:
    BadRequest:
      description: The attempt id is not a UUID, or a registration names an unusable entity, token or provider.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Missing or invalid token.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Valid token for the wrong role.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: No such attempt, or no adapter serves that webhook path.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: The attempt is past `CREATED` and can no longer be cancelled.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unavailable:
      description: The instruments database is unreachable.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Uuid:
      type: string
      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
    NullableUuid:
      type: [string, 'null']
      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
    CurrencyCode:
      type: string
      pattern: '^[A-Z]{3}$'
    AttemptStatus:
      type: string
      description: |
        The attempt's status. The legal moves between these are the D05-5 transition tables, derived from the master's
        state-machine diagrams; nothing changes a status except the transition service, and only along an arrow those
        tables draw.
      enum: [CREATED, SUBMITTING, PENDING, UNKNOWN, SUCCEEDED, DECLINED, FAILED, SETTLED, RETURNED, CANCELLED, NEEDS_REVIEW]
    AttemptDetail:
      type: object
      additionalProperties: false
      required: [attempt, transitions]
      properties:
        attempt: { $ref: '#/components/schemas/Attempt' }
        transitions:
          type: array
          description: Every status change, in `seq` order.
          items: { $ref: '#/components/schemas/AttemptTransition' }
    Attempt:
      type: object
      additionalProperties: false
      # Nullable fields are not required: whether a null is serialized or omitted is a mapper setting, and pinning
      # the contract to it would break on a Jackson default rather than on a real change. additionalProperties still
      # catches a misnamed or invented field, and the schemas below still constrain these when present.
      required: [attempt_id, kind, status, order_group_id, entity_id, provider, currency, amount_minor,
                 blocked_on_capture, version, created_at, updated_at]
      properties:
        attempt_id: { $ref: '#/components/schemas/Uuid' }
        kind:
          type: string
          enum: [CHARGE, REFUND, PAYOUT]
        status: { $ref: '#/components/schemas/AttemptStatus' }
        order_group_id:
          type: string
          description: The trip or group this attempt belongs to; also the partition key of its payment events.
        source_order_id:
          allOf:
            - $ref: '#/components/schemas/NullableUuid'
          description: The COMMERCE order that triggered a charge or refund. Null for payouts, which come from a run.
        payout_run_id:
          allOf:
            - $ref: '#/components/schemas/NullableUuid'
          description: The payout run that created this attempt. Null for charges and refunds.
        entity_id:
          type: string
          description: The rider or driver whose money this moves, for example `rider:R1`.
        provider:
          type: string
          description: The provider this attempt is submitted to, for example `fakecard`.
        currency: { $ref: '#/components/schemas/CurrencyCode' }
        amount_minor:
          type: integer
          format: int64
          minimum: 1
          description: A positive magnitude; the direction is carried by `kind`.
        provider_ref:
          type: [string, 'null']
          description: The provider's own reference, once it has given one.
        failure_code:
          type: [string, 'null']
          description: Why it was declined, failed or returned. Null while the attempt has not failed.
        blocked_on_capture:
          type: boolean
          description: A refund waiting for its original charge to resolve before it can be submitted.
        version:
          type: integer
          format: int64
          minimum: 0
          description: The optimistic guard. Every status change requires the version read with the attempt.
        created_at:
          type: string
          description: RFC 3339 with a Z offset.
        updated_at:
          type: string
          description: RFC 3339 with a Z offset.
    AttemptTransition:
      type: object
      additionalProperties: false
      required: [seq, to_status, cause, at]
      properties:
        seq:
          type: integer
          minimum: 1
          description: Gapless per attempt, starting at 1.
        from_status:
          oneOf:
            - $ref: '#/components/schemas/AttemptStatus'
            - type: 'null'
          description: Null only for a row written without a prior status.
        to_status: { $ref: '#/components/schemas/AttemptStatus' }
        cause:
          type: string
          description: What produced the change, for example a webhook, a sweeper or the cancel endpoint.
        at:
          type: string
          description: RFC 3339 with a Z offset.
...
    AttemptsSummary:
      type: object
      additionalProperties: false
      required: [totals]
      properties:
        totals:
          type: array
          items: { $ref: '#/components/schemas/AttemptStatusTotal' }
    AttemptStatusTotal:
      type: object
      additionalProperties: false
      required: [kind, provider, status, currency, attempts, amount_minor]
      properties:
        kind: { type: string, enum: [CHARGE, REFUND, PAYOUT] }
        provider: { type: string }
        status: { $ref: '#/components/schemas/AttemptStatus' }
        currency: { $ref: '#/components/schemas/CurrencyCode' }
        attempts: { type: integer, format: int64, minimum: 1 }
        amount_minor:
          type: integer
          format: int64
          minimum: 1
          description: The sum of these attempts' positive magnitudes.
...
    PayoutRunListing:
      type: object
      additionalProperties: false
      required: [run_id, currency, status, attempts_created, created_at, outcomes, paid_minor]
      properties:
        run_id: { $ref: '#/components/schemas/Uuid' }
        currency: { $ref: '#/components/schemas/CurrencyCode' }
        status: { type: string, enum: [RUNNING, COMPLETED, REFUSED] }
        refusal_code: { type: [string, 'null'] }
        attempts_created: { type: integer, minimum: 0 }
        created_at: { type: string, description: RFC 3339 with a Z offset. }
        completed_at: { type: [string, 'null'] }
        outcomes:
          type: object
          description: Drivers the run considered, counted by outcome.
          propertyNames:
            enum: [PAID, SKIPPED_IN_FLIGHT, SKIPPED_BELOW_MINIMUM, SKIPPED_NO_BALANCE, SKIPPED_NO_TOKEN]
          additionalProperties: { type: integer, format: int64, minimum: 1 }
        paid_minor:
          type: integer
          format: int64
          minimum: 0
          description: The total of the PAID drivers' amounts, as the ledger reported them when the run read them.
    PayoutRunRequest:
      type: object
      additionalProperties: false
      required: [currency]
      properties:
        currency:
          allOf:
            - $ref: '#/components/schemas/CurrencyCode'
          description: >
            The currency to pay out, from the D01-7 allow-list. Refused rather than normalized: a run keyed on `usd`
            would hash differently from the same run keyed on `USD`, so a caller alternating between them would get
            two runs instead of one replay.
    PayoutRun:
      type: object
      additionalProperties: false
      required: [run_id, currency, status, attempts_created, results, replayed]
      properties:
        run_id: { $ref: '#/components/schemas/Uuid' }
        currency: { $ref: '#/components/schemas/CurrencyCode' }
        status:
          type: string
          enum: [RUNNING, COMPLETED, REFUSED]
          description: >
            `REFUSED` is only ever seen through the 409 problem, never in a 200 or 201 body. `RUNNING` means a run
            that was interrupted between opening and completing; it is returned only by a replay of that key.
        refusal_code:
          type: [string, 'null']
          description: Why a refused run refused — `ledger_stale` or `payouts_disabled`. Null for a completed run.
        attempts_created:
          type: integer
          minimum: 0
          description: How many payout attempts this run actually inserted. Always 0 for a refused run.
        results:
          type: array
          description: Every driver the run considered, ordered by entity id.
          items: { $ref: '#/components/schemas/PayoutRunResult' }
        replayed:
          type: boolean
          description: True when this run already existed and was returned rather than performed again.
    PayoutRunResult:
      type: object
      additionalProperties: false
      required: [entity_id, outcome]
      properties:
        entity_id:
          type: string
          description: The driver, for example `driver:D1`.
        outcome:
          type: string
          enum: [PAID, SKIPPED_IN_FLIGHT, SKIPPED_BELOW_MINIMUM, SKIPPED_NO_BALANCE, SKIPPED_NO_TOKEN]
          description: |
            What happened to this driver, and each value implies a different fix:

            | `outcome` | Meaning |
            |---|---|
            | `PAID` | A payout attempt was created and submitted to the bank. |
            | `SKIPPED_IN_FLIGHT` | The `one_inflight_payout` index refused a second payout; one is already on its way. |
            | `SKIPPED_BELOW_MINIMUM` | The `payable` balance is under the configured minimum, which also covers zero and driver debt. |
            | `SKIPPED_NO_BALANCE` | The ledger holds no entries for this driver at all. |
            | `SKIPPED_NO_TOKEN` | Registered, but at no provider that can disburse. The fix is a registration, not a bigger balance. |
        amount_minor:
          type: [integer, 'null']
          format: int64
          description: >
            What the driver was owed on the account's normal side when the run read it. Null when no balance was
            read, and present but under the minimum for `SKIPPED_BELOW_MINIMUM`.
        as_of_seq:
          type: [integer, 'null']
          format: int64
          description: The ledger sequence the balance was read at, kept so a payout amount can be traced afterwards.
        attempt_id:
          allOf:
            - $ref: '#/components/schemas/NullableUuid'
          description: The payout attempt this run created. Null for every skipped outcome.
    Problem:
      # RFC 9457, plus the stable `code` the master's conventions require.
      type: object
      required: [status, code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code:
          type: string
          # The union of both S05-T11 and S06: every code InstrumentApiException can produce. A code the service
          # can emit but the spec omits is a response its own contract denies, and the conformance test checks it.
          enum: [invalid_attempt_id, attempt_not_found, not_cancellable, invalid_registration, unknown_provider,
                 invalid_signature, invalid_webhook, payload_too_large,
                 # decision: D06-4 — reconciliation (S06-T03)
                 idempotency_key_missing, idempotency_key_reused, invalid_report_date,
                 settlement_reports_unsupported, report_not_ready, provider_unavailable, unsupported_currency,
                 reconciliation_run_not_found,
                 # decision: D05-7, D05-11 — payout runs (S05-T10)
                 ledger_stale, payouts_disabled,
                 # decision: S09 — the dashboard's list endpoints
                 invalid_query,
                 unauthorized, forbidden, database_unavailable]
          description: Stable machine-readable error code; callers branch on this, never on `title`.